# docs/swagger.yaml
definitions:
  response.ErrBodyResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrBodyResponse'
    type: object
  users.GetReviewResponse:
    properties:
      pull_requests:
        items:
          $ref: '#/definitions/users.PullRequestResponse'
        type: array
      user_id:
        type: string
    type: object
  users.PullRequestResponse:
    properties:
      author_id:
        type: string
      pull_request_id:
        type: string
      pull_request_name:
        type: string
      status:
        type: string
    type: object
  users.SetActiveRequest:
    properties:
      is_active:
        type: boolean
      user_id:
        type: string
    type: object
  users.SetIsActiveResponse:
    properties:
      user:
        $ref: '#/definitions/users.UserResponse'
    type: object
  users.UserResponse:
    properties:
      is_active:
        type: boolean
      team_name:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /users/getReview:
    get:
      consumes:
      - application/json
      description: Возвращает список PR'ов, в которых user_id указан как ревьювер
      parameters:
      - description: Идентификатор пользователя
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список PR'ов пользователя
          schema:
            $ref: '#/definitions/users.GetReviewResponse'
        "400":
          description: Некорректный запрос (нет user_id)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получить PR'ы, где пользователь назначен ревьювером
      tags:
      - Users
  /users/setIsActive:
    post:
      consumes:
      - application/json
      description: Принимает user_id и is_active, обновляет пользователя и возвращает
        его состояние
      parameters:
      - description: Тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.SetActiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый пользователь
          schema:
            $ref: '#/definitions/users.SetIsActiveResponse'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Установить флаг активности пользователя
      tags:
      - Users
swagger: "2.0"